kafka and http integrations refactoring to extract common code to AbstractIntegration
